refactor: changed works section to a loop
