+++ public/assets/room/license.txt
