+++ public/assets/room/scene.bin
