+++ public/assets/room/scene.gltf
