+++ public/assets/room/textures/lambert8_baseColor.jpeg
